Fix Sonar findings for TmfEvent and TmfTrace

+++ org.eclipse.linuxtools.tmf.core/src/org/eclipse/linuxtools/tmf/core/trace/TmfCheckpointIndexer.java
@@ -274,7 +274,7 @@ public class TmfCheckpointIndexer<T extends ITmfTrace<ITmfEvent>> implements ITm
             }
         }
         final ITmfContext context = fTrace.seekEvent(location);
-        context.setRank(index * fCheckpointInterval);
+        context.setRank((long) index * fCheckpointInterval);
         return context;
     }
